# archive/idrm-mvp-docs/40-api-openapi.yaml
openapi: 3.0.3
info:
  title: IDRM API
  description: >-
    Integrated Disaster Response Management — MVP API contract.
    Canonical, incident-centric scheme under /api/v1. URIs are a stable contract
    across the SDLC: the FFP adds endpoints at these same paths, never renaming
    MVP ones. Human-readable companion: 40-api-specification.md.
  version: "1.0.0"
  contact:
    name: IDRM Development Team
  license:
    name: Proprietary (government-endorsed)
servers:
  - url: /api/v1
    description: Version-in-path base; host is environment-specific.
...
security:
  - bearerAuth: []
paths:
...
  /auth/register:
    post:
      tags: [auth]
      summary: Register a new account (AC-1.1, AC-1.2)
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/RegisterRequest' }
      responses:
        '201':
          description: Created; verification email sent.
          content:
            application/json:
              schema:
                type: object
                properties:
                  user_id: { type: string, format: uuid }
                  message: { type: string }
        '409': { $ref: '#/components/responses/Conflict' }
        '422': { $ref: '#/components/responses/ValidationError' }
        '429': { $ref: '#/components/responses/RateLimited' }
  /auth/verify-email:
    post:
      tags: [auth]
      summary: Verify email via token
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [token]
              properties: { token: { type: string } }
      responses:
        '200': { $ref: '#/components/responses/Success' }
        '410': { $ref: '#/components/responses/Gone' }
  /auth/resend-verification:
    post:
      tags: [auth]
      summary: Resend verification link
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email]
              properties: { email: { type: string, format: email } }
      responses:
        '200': { $ref: '#/components/responses/Success' }
        '429': { $ref: '#/components/responses/RateLimited' }
  /auth/login:
    post:
      tags: [auth]
      summary: Authenticate and receive tokens (AC-1.3)
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email: { type: string, format: email }
                password: { type: string, format: password }
      responses:
        '200':
          description: Authenticated.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/TokenPair' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '429': { $ref: '#/components/responses/RateLimited' }
  /auth/refresh:
    post:
      tags: [auth]
      summary: Exchange a refresh token for a new access token
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [refresh_token]
              properties: { refresh_token: { type: string } }
      responses:
        '200':
          description: New access token.
          content:
            application/json:
              schema:
                type: object
                properties:
                  access_token: { type: string }
                  token_type: { type: string, example: bearer }
                  expires_in: { type: integer, example: 3600 }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /auth/logout:
    post:
      tags: [auth]
      summary: End the session (AC-1.3)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [refresh_token]
              properties: { refresh_token: { type: string } }
      responses:
        '200': { $ref: '#/components/responses/Success' }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /auth/forgot-password:
    post:
      tags: [auth]
      summary: Request a password-reset link
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email]
              properties: { email: { type: string, format: email } }
      responses:
        '200': { $ref: '#/components/responses/Success' }
        '429': { $ref: '#/components/responses/RateLimited' }
  /auth/reset-password:
    post:
      tags: [auth]
      summary: Reset a password using a token
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [token, new_password]
              properties:
                token: { type: string }
                new_password: { type: string, format: password, minLength: 8 }
      responses:
        '200': { $ref: '#/components/responses/Success' }
        '410': { $ref: '#/components/responses/Gone' }
        '422': { $ref: '#/components/responses/ValidationError' }
  /auth/change-password:
    post:
      tags: [auth]
      summary: Change password while logged in
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [current_password, new_password]
              properties:
                current_password: { type: string, format: password }
                new_password: { type: string, format: password, minLength: 8 }
      responses:
        '200': { $ref: '#/components/responses/Success' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '422': { $ref: '#/components/responses/ValidationError' }

  # ---------------- Users (F1) ----------------
  /users/me:
    get:
      tags: [users]
      summary: Current user profile
      responses:
        '200':
          description: The authenticated user.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/User' }
        '401': { $ref: '#/components/responses/Unauthorized' }
    patch:
      tags: [users]
      summary: Update own profile (email/role immutable here)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name: { type: string }
                phone: { type: string, example: "+919876543210" }
                language: { type: string, enum: [en, hi, te] }
      responses:
        '200':
          description: Updated user.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/User' }
        '422': { $ref: '#/components/responses/ValidationError' }
  /users:
    get:
      tags: [users]
      summary: List/manage users (AC-1.5)
      parameters:
        - { $ref: '#/components/parameters/Page' }
        - { $ref: '#/components/parameters/Limit' }
        - { $ref: '#/components/parameters/Sort' }
        - in: query
          name: role
          schema: { $ref: '#/components/schemas/Role' }
      responses:
        '200':
          description: Paged users.
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/PaginatedEnvelope'
                  - type: object
                    properties:
                      data:
                        type: array
                        items: { $ref: '#/components/schemas/User' }
        '403': { $ref: '#/components/responses/Forbidden' }
  /users/{id}:
    parameters:
      - { $ref: '#/components/parameters/Id' }
    get:
      tags: [users]
      summary: Get a single user (coordinator/admin)
      responses:
        '200':
          description: The user.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/User' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
    patch:
      tags: [users]
      summary: Change a user's role/status (admin only, audited)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                role: { $ref: '#/components/schemas/Role' }
                status: { type: string, enum: [active, suspended] }
      responses:
        '200':
          description: Updated user.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/User' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }

  # ---------------- Incidents (F2–F8) ----------------
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    Id:
      in: path
      name: id
      required: true
      schema: { type: string, format: uuid }
    Page:
      in: query
      name: page
      schema: { type: integer, minimum: 1, default: 1 }
    Limit:
      in: query
      name: limit
      schema: { type: integer, minimum: 1, maximum: 100, default: 20 }
    Sort:
      in: query
      name: sort
      description: "Field to sort by; prefix '-' for descending, e.g. -created_at."
      schema: { type: string, example: -created_at }
...
  responses:
    Success:
      description: Operation succeeded.
      content:
        application/json:
          schema:
            type: object
            properties: { success: { type: boolean, example: true } }
    Unauthorized:
      description: Missing or invalid authentication.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    Forbidden:
      description: Authenticated but not permitted.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    NotFound:
      description: Resource not found.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    Conflict:
      description: State or uniqueness conflict.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    Gone:
      description: Token expired / resource gone.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    ValidationError:
      description: Semantic validation failed.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    RateLimited:
      description: Rate limit exceeded.
      headers:
        Retry-After:
          schema: { type: integer }
          description: Seconds to wait before retrying.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
...
  schemas:
    Role:
      type: string
      enum: [citizen, provider, coordinator, admin]
...
    Pagination:
      type: object
      properties:
        page: { type: integer }
        limit: { type: integer }
        total: { type: integer }
        total_pages: { type: integer }
    PaginatedEnvelope:
      type: object
      properties:
        data: { type: array, items: {} }
        pagination: { $ref: '#/components/schemas/Pagination' }
    Error:
      type: object
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code: { type: string, example: already_accepted }
            message: { type: string }
            details:
              type: array
              items:
                type: object
                properties:
                  field: { type: string }
                  issue: { type: string }
    RegisterRequest:
      type: object
      required: [email, password, name, phone, role]
      properties:
        email: { type: string, format: email }
        password: { type: string, format: password, minLength: 8 }
        name: { type: string }
        phone: { type: string, example: "+919876543210" }
        role: { $ref: '#/components/schemas/Role' }
    TokenPair:
      type: object
      properties:
        access_token: { type: string }
        refresh_token: { type: string }
        token_type: { type: string, example: bearer }
        expires_in: { type: integer, example: 3600 }
        user: { $ref: '#/components/schemas/User' }
    User:
      type: object
      properties:
        id: { type: string, format: uuid }
        email: { type: string, format: email }
        name: { type: string }
        phone: { type: string }
        role: { $ref: '#/components/schemas/Role' }
        status: { type: string, enum: [pending, active, suspended] }
        language: { type: string, enum: [en, hi, te] }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
...